Created Sidebar component
