+++ app/AdminPortal/StudentCourses/page.js
@@ -1,7 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import Header from "../components/header";
+import Header from "../components/Header";
 // import Sidebar from "../components/sidebar";
 import styles from "./studentCourses.module.css";
