+++ app/AdminPortal/components/Sidebar.module.css
@@ -1,71 +1,140 @@
-.sidebar {
-  width: 250px;
-  color: black;
-  padding: 20px;
-  position: fixed;
-  height: calc(100% - 140px); /* Subtract header height (adjust as needed) */
-  top: 140px; /* Header height */
-  left: 0;
-  z-index: 1000; /* Ensure it's above other elements */
-}
-
-.sidebar {
-  box-shadow: 3px 0 5px rgba(0, 0, 0, 0.2);
-}
-.sidebar .logo {
-  text-align: center;
-  margin-bottom: 20px;
-  margin-left: 20px;
-}
-
-.sidebar h1 {
-  font-size: 20px;
-  margin-top: 10px;
-  color: white;
-  margin-right: 50px;
-}
-
-.navbar {
+/* 📌 Sidebar Wrapper - Positioned Below Header */
+.sidebarWrapper {
   display: flex;
-  flex-direction: column;
+  flex-direction: row;
+  width: 100%;
+  height: calc(100vh - 60px); /* Adjust for header height */
+  position: relative;
 }
 
-.navbar a {
-  color: #005f73;
-  text-decoration: none;
-  margin: 10px 0;
-  padding: 8px;
-  border-radius: 4px;
-  transition: background-color 0.3s;
-}
-
-.navbar a:hover {
-  background-color: rgba(117, 208, 236, 0.2);
-  transform: translateY(-3px);
-}
-
-.studentInfo {
-  margin-bottom: 20px;
-}
-
-.logoutButton {
-  background-color: #ff4d4f;
+/* Sidebar Styles */
+.sidebar {
+  position: fixed;
+  top: 60px; /* ✅ Now Positioned Beneath the Header */
+  left: 0;
+  width: 250px;
+  height: calc(100vh - 60px);
+  background-color: #1a202c;
   color: white;
+  transition: width 0.3s ease-in-out;
+  overflow-x: hidden;
+  box-shadow: 2px 0px 10px rgba(0, 0, 0, 0.2);
+}
+
+/* Sidebar Collapsed */
+.sidebar.closed {
+  width: 70px;
+}
+
+/* Sidebar Header */
+.sidebarHeader {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 15px;
+  background-color: #008080;
+}
+
+/* Sidebar Toggle Button */
+.sidebarToggle {
+  background: none;
   border: none;
-  padding: 10px 15px;
-  border-radius: 5px;
+  color: white;
+  font-size: 20px;
   cursor: pointer;
-  margin-top: 20px;
-  font-size: 16px;
-  transition: background-color 0.3s;
 }
 
-.logoutButton:hover {
-  background-color: #e63946;
+/* Sidebar Menu */
+.sidebarMenu {
+  padding: 10px;
 }
 
-/* Ensure main content is not overlapped by the sidebar */
-.main-content {
-  margin-left: 250px; /* Prevent content from being hidden under the sidebar */
-  padding: 20px;
+/* Sidebar Item */
+.sidebarItem {
+  cursor: pointer;
+  position: relative;
+  transition: background 0.3s;
+}
+
+/* Sidebar Title */
+.sidebarTitle {
+  display: flex;
+  align-items: center;
+  padding: 12px;
+  border-radius: 5px;
+  transition: background 0.2s;
+}
+
+.sidebarIcon {
+  margin-right: 12px;
+}
+
+.sidebarTitle:hover {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+/* Sidebar Chevron (Dropdown Icon) */
+.sidebarChevron {
+  margin-left: auto;
+}
+
+/* Sidebar Submenu */
+.sidebarSubmenu {
+  padding-left: 20px;
+  display: none;
+}
+
+/* ✅ Submenu only visible when sidebar is open */
+.sidebar.open .sidebarSubmenu {
+  display: block;
+}
+
+/* ❌ Hide submenu completely when sidebar is collapsed */
+.sidebar.closed .sidebarSubmenu {
+  display: none !important;
+}
+
+/* Sidebar Subitem */
+.sidebarSubitem {
+  padding: 8px;
+  font-size: 14px;
+  color: white;
+  transition: background 0.2s;
+}
+
+.sidebarSubitem:hover {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+/* ✅ Main Content Adjusts When Sidebar Expands */
+.mainContent {
+  margin-left: 250px;
+  width: calc(100% - 250px);
+  min-height: 100vh;
+  background-color: #f4f6f9;
+  padding: 80px 20px 20px;
+}
+
+/* ✅ Closed Sidebar Main Content */
+.sidebar.closed + .mainContent {
+  margin-left: 70px;
+  width: calc(100% - 70px);
+}
+
+/* ✅ Prevent Black Space */
+body {
+  margin: 0;
+  padding: 0;
+}
+
+/* Responsive Fixes */
+@media screen and (max-width: 768px) {
+  .sidebar {
+      width: 70px;
+  }
+
+  .mainContent {
+      margin-left: 70px;
+      width: calc(100% - 70px);
+  }
 }
